+++ tsdown.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   entry: "src/index.ts",
   exports: true,
   deps: {
-    onlyAllowBundle: ["@types/geojson"],
+    onlyBundle: ["@types/geojson"],
   },
 })